Fix defect where invalid tag causes exception
Removing assert that raises exception when no posts
are found with a specified tag.  This is caused when
future posts contain new tags.  This is a quick fix
to no longer cause an exception, the underlying issue
is that the future tags should not be shown on the page
to begin with.  That fix will be done in the future at
which time this assert can be readded.